Various improvements to codebase.
1. Moved main terminal handling to functions provided in term.py
This creates a cleaner, easier-to-work-with codebase because there
won't be as much clutter in Emacs.

2. Moved dialogue and whatnot to dialog.py
This allows me to easily change dialogue in the future without
messing in the codebase too much. And also if I ever choose to
translate this to English again then it shouldn't be too hard.

3. Replace string with lists for dialogue
This allows me to print whatever can be printed without it crashing.

Signed-off-by: Abdul Karim Kikar <abdul@oklomsy.com>