Working on cell generation
Everything seems good! I just need to figure out how to track
the columns and rows!

I could use a variable to store it's previous value when it does get
iterated and then use a IF condition to see if it has changed on each print
in the for loop (in actualview())

The more important part would be tracking columns! rather than rows.
And instead of blindly printing, we could make it so after each cell
getting printed, we add one to the col variable and then we add a 1 to
row when col surpasses lib.collimit or similar.

Signed-off-by: Abdul Karim Kikar <abdu1771@appsodense.dk>