Why doesn't it work?
So I have a problem with the parser, when checking if a character exists, it just doesn't work.
But if I try it out with the regular Python REPL prompt then it DOES work.

What did I do to deserve this mysterious ghost bug?!?

I don't have that many days left! And I haven't done a lot!
I better increase my effort by 200x or I will fail this!

Signed-off-by: Abdul Karim Kikar <abdu1771@appsodense.dk>