Ok I got a bit far, I implemented dayHasEvent()
The problem is that the calendar takes about 100ms to open
it has a bit of visible lag but my processor is stong enough
to cope.

If this ever were to turn into a real product then some more
optimizations would have to be done and maybe even a rewrite of
the entire codebase, not because it's bad but because it's a teensy bit
slow.

Signed-off-by: Abdul Karim Kikar <abdu1771@appsodense.dk>