Moved the basic help dialog of calview() to dialog.py
There is still quite a lot of stuff remaining but I can
always work on this in the weekends with a nice cup of tea.

or Joe, if you really, really hate Joe. :P

Alright, tomorrow, I will work on the smart weather function
Among other things. I will also need to work on the following:
1. Switch to next and previous months with up and down arrow.
2. Editing functionality (This could be in a separate program,
   caledit.py)
3. Fixing all the inevitable bugs :)

Signed-off-by: Abdul Karim Kikar <abdu1771@appsodense.dk>